% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSigmaS_root.R
\name{generateSigmaS_root}
\alias{generateSigmaS_root}
\title{Posterior for sigma (generateSigmaS_root)}
\usage{
generateSigmaS_root(A, Z, U, N, asigma, bsigma, D, IND)
}
\arguments{
\item{A}{array \cr
The sufficient statistic A.}

\item{Z}{array \cr
The sufficient statistic Z.}

\item{U}{array \cr
The simulated values of U during the iteration.}

\item{N}{int \cr
The number of data.}

\item{asigma}{double \cr
The shape parameter of the Gamma prior.}

\item{bsigma}{double \cr
The rate parameter of the Gamma prior.}

\item{D}{int \cr
The dimension of data.}

\item{IND}{int_array \cr
The array with the accepted principal axes.}
}
\value{
\code{generateSigmaS_root} returns a value:
\item{sigmaS}{
                array \cr
                the simulation from the full conditional posterior of
                Sigma.
          }
}
\description{
Generate Sigma from the full conditional posterior.
}
\details{
Using the full conditional posterior computed in the paper, we simulate
Sigma, which will be used for the computations of variance-covariance matrix
of our data.
}
\note{
We remark that only the activated principal axes are used to
      obtain a simulated sample. Hence we set as default value for the
      rejected one the last iterations used.
}
\references{
\itemize{

 \item   [1] Y. Wang, A. Canale, D. Dunson.
         "Scalable Geometric Density Estimation" (2016).\cr
         The implementation of rgeode_root is inspired to the
         Matlab implementation of generateSigmaS_root by Ye Wang.
}
}
\author{
L. Rimella, \email{lorenzo.rimella@hotmail.it}
}
