% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEODE_impute.R
\name{GEODE_impute}
\alias{GEODE_impute}
\title{Missing Data Imputation (GEODE_impute)}
\usage{
GEODE_impute(U, SIGMAS, LAMBDA, MU, AA, Z, id_m, pos_m, IND)
}
\arguments{
\item{U}{array \cr
The estimation for u.}

\item{SIGMAS}{double \cr
The estimation for sigma.}

\item{LAMBDA}{matrix \cr
The matrix W.}

\item{MU}{array \cr
The mean.}

\item{AA}{array \cr
The sufficient statistic A.}

\item{Z}{matrix \cr
The sufficient statistic Z.}

\item{id_m}{array \cr
The array with the rows containing missing data.}

\item{pos_m}{list \cr
The list containing the position of the missing data
for each row.}

\item{IND}{array \cr
The set of accepted Principal Axes.}
}
\value{
\code{GEODE_impute} returns a value:
\item{y_m}{     list \cr
                A list containing the simulated Missing Data.
        }
}
\description{
Generate pseudo-observation from a Multivariate Normal
Distribution to substitute Missing Data.
}
\references{
\itemize{

 \item   [1] Y. Wang, A. Canale, D. Dunson.
         "Scalable Geometric Density Estimation" (2016).\cr
         The implementation of GEODE_impute is inspired to the
         Matlab implementation of GEODE_impute by Ye Wang.
}
}
\author{
L. Rimella, \email{lorenzo.rimella@hotmail.it}
}
