% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeode_root_m.R
\name{rgeode_root_m}
\alias{rgeode_root_m}
\title{The core of the Gibbs sampling with missing data.}
\usage{
rgeode_root_m(Y, d, burn, its, tol, atau, asigma, bsigma, starttime, stoptime,
  fast, c0, c1)
}
\arguments{
\item{Y}{array_like \cr
a real input matrix (or data frame), with dimensions
\eqn{(n, D)}. It is the real matrix of data.}

\item{d}{int, optional \cr
it is the conservative upper bound for the dimension
D. We are confident that the real dimension is smaller
then it.}

\item{burn}{int, optional \cr
number of burn-in to perform in our Gibbs sampler. It
represents also the stopping time that stop the choice
of the principal axes.}

\item{its}{int, optional \cr
number of iterations that must be performed after
the burn-in.}

\item{tol}{double, optional \cr
threshold for adaptively removing redundant
dimensions. It is used compared with the ratio:
\eqn{\frac{\alpha_j^2(t)}{\max \alpha_i^2(t)}}.}

\item{atau}{double, optional \cr
The parameter \eqn{a_\tau} of the truncated
Exponential (the prior for \eqn{\tau_j}).}

\item{asigma}{double, optional \cr
The shape parameter \eqn{a_\sigma} of the
truncated Gamma (the prior for \eqn{\sigma^2}).}

\item{bsigma}{double, optional \cr
The rate parameter \eqn{b_\sigma} of the
truncated Gamma (the prior for \eqn{\sigma^2}).}

\item{starttime}{int, optional \cr
starting time for adaptive pruning. It must be less
then the number of burn-in.}

\item{stoptime}{int, optional \cr
stop time for adaptive pruning. It must be less
then the number of burn-in.}

\item{fast}{bool, optional \cr
If \eqn{TRUE} it is run using fast d-rank SVD.
Otherwise it uses the classical SVD.}

\item{c0}{double, optional \cr
Additive constant for the exponent of the pruning step.}

\item{c1}{double, optional \cr
Multiplicative constant for the exponent of the pruning step.}
}
\value{
The outputs are the same as in \code{rgeode}.
}
\description{
Perform GEODE with missing data.
}
\note{
The parameters are the same as in \code{rgeode}.
}
\references{
\itemize{

 \item   [1] Y. Wang, A. Canale, D. Dunson.
         "Scalable Geometric Density Estimation" (2016).\cr
         The implementation of rgeode_root is inspired to the
         Matlab implementation of GEODE_root by Ye Wang.
}
}
\author{
L. Rimella, \email{lorenzo.rimella@hotmail.it}
}
